refactor: Introduce <Calculator> component
